Rename src/os/linux to src/os/unix
