+++ CodeBlocks/os/os.cbp
@@ -384,40 +384,40 @@
 		<Unit filename="../../src/os/fmtspecs_cvt.h" />
 		<Unit filename="../../src/os/funcwrap.h" />
 		<Unit filename="../../src/os/lib.h" />
-		<Unit filename="../../src/os/linux/KeySend.cpp" />
-		<Unit filename="../../src/os/linux/PWYubi.cpp">
+		<Unit filename="../../src/os/unix/KeySend.cpp" />
+		<Unit filename="../../src/os/unix/PWYubi.cpp">
 			<Option target="Debug_L64Y" />
 			<Option target="Debug_L32Y" />
 			<Option target="Release_L64Y" />
 			<Option target="Release_L32Y" />
 		</Unit>
-		<Unit filename="../../src/os/linux/PWYubi.h">
+		<Unit filename="../../src/os/unix/PWYubi.h">
 			<Option target="Debug_L64Y" />
 			<Option target="Debug_L32Y" />
 			<Option target="Release_L64Y" />
 			<Option target="Release_L32Y" />
 		</Unit>
-		<Unit filename="../../src/os/linux/UUID.cpp" />
-		<Unit filename="../../src/os/linux/debug.cpp" />
-		<Unit filename="../../src/os/linux/dir.cpp" />
-		<Unit filename="../../src/os/linux/env.cpp" />
-		<Unit filename="../../src/os/linux/file.cpp" />
-		<Unit filename="../../src/os/linux/logit.cpp" />
-		<Unit filename="../../src/os/linux/media.cpp" />
-		<Unit filename="../../src/os/linux/mem.cpp" />
-		<Unit filename="../../src/os/linux/pws_str.cpp" />
-		<Unit filename="../../src/os/linux/pws_str.h" />
-		<Unit filename="../../src/os/linux/pws_time.cpp" />
-		<Unit filename="../../src/os/linux/pws_time.h" />
-		<Unit filename="../../src/os/linux/rand.cpp" />
-		<Unit filename="../../src/os/linux/registry.cpp" />
-		<Unit filename="../../src/os/linux/run.cpp" />
-		<Unit filename="../../src/os/linux/sleep.cpp" />
-		<Unit filename="../../src/os/linux/unicode2keysym.cpp" />
-		<Unit filename="../../src/os/linux/unicode2keysym.h" />
-		<Unit filename="../../src/os/linux/utf8conv.cpp" />
-		<Unit filename="../../src/os/linux/xsendstring.cpp" />
-		<Unit filename="../../src/os/linux/xsendstring.h" />
+		<Unit filename="../../src/os/unix/UUID.cpp" />
+		<Unit filename="../../src/os/unix/debug.cpp" />
+		<Unit filename="../../src/os/unix/dir.cpp" />
+		<Unit filename="../../src/os/unix/env.cpp" />
+		<Unit filename="../../src/os/unix/file.cpp" />
+		<Unit filename="../../src/os/unix/logit.cpp" />
+		<Unit filename="../../src/os/unix/media.cpp" />
+		<Unit filename="../../src/os/unix/mem.cpp" />
+		<Unit filename="../../src/os/unix/pws_str.cpp" />
+		<Unit filename="../../src/os/unix/pws_str.h" />
+		<Unit filename="../../src/os/unix/pws_time.cpp" />
+		<Unit filename="../../src/os/unix/pws_time.h" />
+		<Unit filename="../../src/os/unix/rand.cpp" />
+		<Unit filename="../../src/os/unix/registry.cpp" />
+		<Unit filename="../../src/os/unix/run.cpp" />
+		<Unit filename="../../src/os/unix/sleep.cpp" />
+		<Unit filename="../../src/os/unix/unicode2keysym.cpp" />
+		<Unit filename="../../src/os/unix/unicode2keysym.h" />
+		<Unit filename="../../src/os/unix/utf8conv.cpp" />
+		<Unit filename="../../src/os/unix/xsendstring.cpp" />
+		<Unit filename="../../src/os/unix/xsendstring.h" />
 		<Unit filename="../../src/os/logit.h" />
 		<Unit filename="../../src/os/media.h" />
 		<Unit filename="../../src/os/mem.h" />
