+++ CodeLite/os.project
@@ -30,29 +30,29 @@
   <Dependencies/>
   <VirtualDirectory Name="src"/>
   <VirtualDirectory Name="include"/>
-  <VirtualDirectory Name="linux">
-    <File Name="../src/os/linux/sleep.cpp"/>
-    <File Name="../src/os/linux/pws_time.cpp"/>
-    <File Name="../src/os/linux/xsendstring.cpp"/>
-    <File Name="../src/os/linux/mem.cpp"/>
-    <File Name="../src/os/linux/xsendstring.h"/>
-    <File Name="../src/os/linux/utf8conv.cpp"/>
-    <File Name="../src/os/linux/KeySend.cpp"/>
-    <File Name="../src/os/linux/env.cpp"/>
-    <File Name="../src/os/linux/pws_str.h"/>
-    <File Name="../src/os/linux/file.cpp"/>
-    <File Name="../src/os/linux/run.cpp"/>
-    <File Name="../src/os/linux/pws_time.h"/>
-    <File Name="../src/os/linux/debug.cpp"/>
-    <File Name="../src/os/linux/registry.cpp"/>
-    <File Name="../src/os/linux/rand.cpp"/>
-    <File Name="../src/os/linux/pws_str.cpp"/>
-    <File Name="../src/os/linux/dir.cpp"/>
-    <File Name="../src/os/linux/UUID.cpp"/>
-    <File Name="../src/os/linux/unicode2keysym.cpp"/>
-    <File Name="../src/os/linux/unicode2keysym.h"/>
-    <File Name="../src/os/linux/logit.cpp"/>
-    <File Name="../src/os/linux/media.cpp"/>
+  <VirtualDirectory Name="unix">
+    <File Name="../src/os/unix/sleep.cpp"/>
+    <File Name="../src/os/unix/pws_time.cpp"/>
+    <File Name="../src/os/unix/xsendstring.cpp"/>
+    <File Name="../src/os/unix/mem.cpp"/>
+    <File Name="../src/os/unix/xsendstring.h"/>
+    <File Name="../src/os/unix/utf8conv.cpp"/>
+    <File Name="../src/os/unix/KeySend.cpp"/>
+    <File Name="../src/os/unix/env.cpp"/>
+    <File Name="../src/os/unix/pws_str.h"/>
+    <File Name="../src/os/unix/file.cpp"/>
+    <File Name="../src/os/unix/run.cpp"/>
+    <File Name="../src/os/unix/pws_time.h"/>
+    <File Name="../src/os/unix/debug.cpp"/>
+    <File Name="../src/os/unix/registry.cpp"/>
+    <File Name="../src/os/unix/rand.cpp"/>
+    <File Name="../src/os/unix/pws_str.cpp"/>
+    <File Name="../src/os/unix/dir.cpp"/>
+    <File Name="../src/os/unix/UUID.cpp"/>
+    <File Name="../src/os/unix/unicode2keysym.cpp"/>
+    <File Name="../src/os/unix/unicode2keysym.h"/>
+    <File Name="../src/os/unix/logit.cpp"/>
+    <File Name="../src/os/unix/media.cpp"/>
   </VirtualDirectory>
   <VirtualDirectory Name="os">
     <File Name="../src/os/env.h"/>
