+++ Makefile.linux
@@ -47,7 +47,7 @@ export CPPFLAGS += -std=c++11
 all: I18N unicodedebug unicoderelease
 
 unicodedebug unicoderelease:
-	$(MAKE) -C src/os/linux $@
+	$(MAKE) -C src/os/unix $@
 	$(MAKE) -C src/core $@
 	$(MAKE) -C src/ui/wxWidgets CONFIG=$@
 
@@ -61,13 +61,13 @@ release:
 clean: debug-clean release-clean
 
 debug-clean:
-	$(MAKE) CONFIG=unicodedebug -C src/os/linux clean
+	$(MAKE) CONFIG=unicodedebug -C src/os/unix clean
 	$(MAKE) CONFIG=unicodedebug -C src/core clean
 	$(MAKE) CONFIG=unicodedebug -C src/ui/wxWidgets clean
 	$(MAKE) CONFIG=unicodedebug -C src/test clean
 
 release-clean:
-	$(MAKE) CONFIG=unicoderelease -C src/os/linux clean
+	$(MAKE) CONFIG=unicoderelease -C src/os/unix clean
 	$(MAKE) CONFIG=unicoderelease -C src/core clean
 	$(MAKE) CONFIG=unicoderelease -C src/ui/wxWidgets clean
 	$(MAKE) CONFIG=unicoderelease -C src/test clean
