+++ src/os/CMakeLists.txt
@@ -51,26 +51,26 @@ endif ()
     )
   elseif (${CMAKE_SYSTEM_NAME} MATCHES "Linux")
   set (OS_SRCS
-    linux/KeySend.cpp
-    linux/UUID.cpp
-    linux/debug.cpp
-    linux/dir.cpp
-    linux/env.cpp
-    linux/file.cpp
-    linux/logit.cpp
-    linux/media.cpp
-    linux/mem.cpp
-    linux/pws_str.cpp
-    linux/pws_time.cpp
-    linux/rand.cpp
-    linux/registry.cpp
-    linux/run.cpp
-    linux/sleep.cpp
-    linux/unicode2keysym.cpp
-    linux/utf8conv.cpp
-    linux/xsendstring.cpp
+    unix/KeySend.cpp
+    unix/UUID.cpp
+    unix/debug.cpp
+    unix/dir.cpp
+    unix/env.cpp
+    unix/file.cpp
+    unix/logit.cpp
+    unix/media.cpp
+    unix/mem.cpp
+    unix/pws_str.cpp
+    unix/pws_time.cpp
+    unix/rand.cpp
+    unix/registry.cpp
+    unix/run.cpp
+    unix/sleep.cpp
+    unix/unicode2keysym.cpp
+    unix/utf8conv.cpp
+    unix/xsendstring.cpp
     )
-  set (OS_YUBI_SRC linux/PWYubi.cpp)
+  set (OS_YUBI_SRC unix/PWYubi.cpp)
 else()
   message (FATAL_ERROR "Unsupported OS "
     ${CMAKE_SYSTEM_NAME}
