+++ src/os/typedefs.h
@@ -153,7 +153,7 @@ typedef int HANDLE;
 
 /* These two files require the above definitions */
 #include "debug.h"
-#include "linux/pws_time.h"
+#include "unix/pws_time.h"
 #endif /* _WIN32 */
 
 // Compile-time check that our sized types are correct:
