+++ src/os/unix/rand.cpp
