+++ src/os/unix/registry.cpp
