+++ src/os/unix/xsendstring.cpp
