+++ src/os/unix/xsendstring.h
