+++ src/ui/wxWidgets/I18N/CMakeLists.txt
@@ -28,8 +28,8 @@ file (GLOB I18N_SRCS
   "${SRC_BASE_DIR}/ui/wxWidgets/*.h"
   "${SRC_BASE_DIR}/core/*.cpp"
   "${SRC_BASE_DIR}/core/*.h"
-  "${SRC_BASE_DIR}/os/linux/*.cpp"
-  "${SRC_BASE_DIR}/os/linux/*.h"
+  "${SRC_BASE_DIR}/os/unix/*.cpp"
+  "${SRC_BASE_DIR}/os/unix/*.h"
   )
 
 # pot is updated from source code
